BACKPORT: net: Introduce parse_protocol header_ops callback
Introduce a new optional header_ops callback called parse_protocol and a
wrapper function dev_parse_header_protocol, similar to dev_parse_header.

The new callback's purpose is to extract the protocol number from the L2
header, the format of which is known to the driver, but not to the upper
layers of the stack.

Change-Id: Idbb518f4580c35703e7da015fb4615fb7023677d

Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>