mm: kmemleak: Don't die when memory allocation fails
When memory is leaking, it's going to be harder to allocate more memory,
making it more likely for this failure condition inside of kmemleak to
manifest itself. This is extremely frustrating since kmemleak kills
itself upon the first instance of memory allocation failure.

Bypass that and make kmemleak more resilient when memory is running low.

Change-Id: I17ce7737b7aa3f2e8cfada74ba5a94ace8114858

Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>