UPSTREAM: wireguard: socket: ignore v6 endpoints when ipv6 is disabled
The previous commit fixed a memory leak on the send path in the event
that IPv6 is disabled at compile time, but how did a packet even arrive
there to begin with? It turns out we have previously allowed IPv6
endpoints even when IPv6 support is disabled at compile time. This is
awkward and inconsistent. Instead, let's just ignore all things IPv6,
the same way we do other malformed endpoints, in the case where IPv6 is
disabled.

Fixes: e7096c131e51 ("net: WireGuard secure network tunnel")
[ Tashar02: Upstream to mainline ]

Change-Id: I5be677ecee2367e4fde7f0c2835d6b67ba1241d6

Signed-off-by: nerd.zenss_ <nerd.zensdev@gmail.com>